+++ packages/core/src/formatters/math/even.formatter.ts
@@ -1,5 +1,4 @@
 /**
- * Prüft ob eine Zahl gerade ist oder nicht
  * Check if a number is even or not
  */
 export const even = {