+++ packages/shopify/src/formatters/img-url.formatter.ts
@@ -1,6 +1,4 @@
-/* tslint:disable:variable-name */
-
-import { ShopifyService } from '../services/shopify.service';
+import { IFormatter } from '@ribajs/core';
 
 /**
  * Return a resized shopify image URL
@@ -12,4 +10,25 @@ import { ShopifyService } from '../services/shopify.service';
  * @param crop TODO
  * @param extension
  */
-export const img_url = ShopifyService.resizeImage;
+export const imgUrlFormatter: IFormatter = {
+  name: 'img_url',
+  read(url: string, size: string, scale: number, crop: string, extension: string) {
+    try {
+      if ('original' === size || 'master' === size) {
+        return url;
+      }
+      const result = url.match(/(.*\/[\w\-\_\.]+)\.(\w{2,4})/);
+
+      if (!result || !result[1] || !result[2]) {
+        throw new Error(`Can't match url ${url}`);
+      }
+
+      const path = result[1];
+      extension = extension || result[2];
+      return path + '_' + size + '.' + extension;
+    } catch (error) {
+      console.error(error);
+      return url;
+    }
+  },
+};