+++ packages/shopify/src/formatters/money-with-currency.formatter.ts
@@ -1,11 +1,19 @@
-/* tslint:disable:variable-name */
-
+import { IFormatter } from '@ribajs/core';
+import { moneyFormatter } from './money.formatter';
 import { ShopifyService } from '../services/shopify.service';
-
-const shopifyService = new ShopifyService();
 
 /**
  * Formats the price based on the shop's HTML with currency setting (if the format is not overwritten by passing a format parameter).
  * @see https://help.shopify.com/en/themes/liquid/filters/money-filters
  */
-export const money_with_currency = shopifyService.formatMoneyWithCurrency;
+export const moneyWithCurrencyFormatter: IFormatter = {
+  name: 'money_with_currency',
+  read(cents: string | number, format?: string) {
+    const formatString = format || ShopifyService.moneyWithCurrencyFormat;
+    if (!moneyFormatter.read) {
+      console.error(new Error('Can\'t find moneyFormatter\'s read method!'));
+      return cents;
+    }
+    return moneyFormatter.read(cents, formatString);
+  },
+};