+++ packages/shopify/src/index.ts
@@ -1,7 +1,5 @@
 import { IRibaModule } from '@ribajs/core';
 
-export * from './components';
-export * from './formatters';
 export * from './services';
 export * from './interfaces';